+++ pypsdm/models/input/entity.py
@@ -158,13 +158,6 @@ class Entities(ABC):
         """
         return self.data["operator"]
 
-    @property
-    @abstractmethod
-    def node(self) -> Series:
-        """
-        Returns: The nodes to which the entities are connected.
-        """
-
     def get(self, uuid: str) -> Series:
         """
         Returns the entity information of the entitiy with the given uuid.
@@ -324,8 +317,8 @@ class Entities(ABC):
     ) -> EntityType:
         """
         Creates a copy of the current Entities instance.
-        By default does a deep copy of all data and replaces the given changes.
-        When deep is false, only the references to the data of the non-changed attribtues are copied.
+        By default, does a deep copy of all data and replaces the given changes.
+        When deep is false, only the references to the data of the non-changed attributes are copied.
 
         Args:
             deep: Whether to do a deep copy of the data.
@@ -350,7 +343,7 @@ class Entities(ABC):
         Args:
             path: The path to the csv file.
             delimiter: The delimiter of the csv file.
-            must_exist: Wether or not exception is thrown if file does not exist.
+            must_exist: Whether exception is thrown if file does not exist.
 
         Returns:
            The corresponding entities object.
@@ -459,8 +452,8 @@ class Entities(ABC):
         Returns the corresponding entity enum value.
         """
 
-    @classmethod
-    def attributes(cls) -> list[str]:
+    @staticmethod
+    def attributes() -> list[str]:
         """
         Method that should hold all attributes field (transformed to snake_case and case-sensitive)
         of the corresponding PSDM entity