+++ pypsdm/models/input/participant/em.py
@@ -1,11 +1,14 @@
 from dataclasses import dataclass
 
+from pandas import Series
+
 from pypsdm.models.enums import EntitiesEnum, SystemParticipantsEnum
-from pypsdm.models.input.participant.participant import SystemParticipants
+from pypsdm.models.input.entity import Entities
 
 
 @dataclass(frozen=True)
-class EnergyManagementSystems(SystemParticipants):
+class EnergyManagementSystems(Entities):
+
     def __eq__(self, other: object) -> bool:
         return super().__eq__(other)
 
@@ -14,16 +17,16 @@ class EnergyManagementSystems(SystemParticipants):
         return SystemParticipantsEnum.ENERGY_MANAGEMENT
 
     @property
-    def control_strategy(self):
+    def control_strategy(self) -> Series:
         return self.data["control_strategy"]
 
     @property
-    def parent_em(self):
+    def parent_em(self) -> Series:
         return self.data["parent_em"]
 
     @staticmethod
-    def attributes():
-        return SystemParticipants.attributes() + [
+    def attributes() -> list[str]:
+        return Entities.attributes() + [
             "control_strategy",
             "parent_em",
         ]