+++ pypsdm/models/result/container/raw_grid.py
@@ -80,7 +80,7 @@ class RawGridResultContainer(ResultContainerMixin):
 
         NOTE: This only makes sense if the entities indexes are continuous. Given
         that we deal with discrete event data that means that the last state of self
-        is valid until the first state of other. Which would probably not be what
+        is valid until the first state of the other. Which would probably not be what
         you want in case the results are separated by a year.
 
         Args: